Add comments on Files and URLs

Add comments on files and URLs. For instance, flagging false positives, adding details about malware, instructions for cleaning malware, etc.

add_comments(hash = NULL, comment = NULL, ...)

Arguments

hash

hash for the resource you want to comment on; Required; String

comment

review; Required; String

...

Additional arguments passed to virustotal2_POST.

Value

data.frame with 2 columns: response_code, verbose_msg

If the hash is incorrect or a duplicate comment is posted, response_code will be 0

If the hash is incorrect, verbose_msg will be 'Invalid resource'

If a duplicate comment is posted, verbose_msg will be 'Duplicate comment'

If a comment is posted successfully, response_code will be 1 and verbose_msg will be 'Your comment was successfully posted'

References

https://developers.virustotal.com/v2.0/reference

See also

Examples

if (FALSE) { # \dontrun{

# Before calling the function, set the API key using set_key('api_key_here')
  
add_comments(hash='99017f6eebbac24f351415dd410d522d', comment="This is great.")

} # }